Preserve source paths for Nix modules
In order for the module system to track the source location of modules,
they must not be imported and instead just passed to imports as paths.
This commit adds the ability to mark config values such that if those
values are are loaded from `nixDir` and loaded as an attrset, then the
values are just set as paths to the nix files instead of importing them.